Move RSlotStack from RDF to Imt
