+++ tree/dataframe/test/dataframe_nodes.cxx
@@ -1,7 +1,7 @@
 #include "ROOT/TestSupport.hxx"
 
 #include <ROOT/RDataFrame.hxx>
-#include <ROOT/RDF/RSlotStack.hxx>
+#include <ROOT/RSlotStack.hxx>
 #include <TStatistic.h> // To check reading of columns with types which are mothers of the column type
 #include <TSystem.h>
 
@@ -16,7 +16,7 @@ TEST(RDataFrameNodes, RSlotStackGetOneTooMuch)
 {
    auto theTest = []() {
       unsigned int n(2);
-      ROOT::Internal::RDF::RSlotStack s(n);
+      ROOT::Internal::RSlotStack s(n);
 
       std::vector<std::thread> ts;
 
@@ -34,7 +34,7 @@ TEST(RDataFrameNodes, RSlotStackGetOneTooMuch)
 TEST(RDataFrameNodes, RSlotStackPutBackTooMany)
 {
    auto theTest = []() {
-      ROOT::Internal::RDF::RSlotStack s(1);
+      ROOT::Internal::RSlotStack s(1);
       s.ReturnSlot(0);
    };
